requirements.txt
Django==2.2
djangorestframework>=3.11.0
django-filter
djangorestframework-simplejwt
django-cors-headers